# src/interface/http/v1/internal/schemas.py
from __future__ import annotations

from pydantic import BaseModel


class MoneyResponse(BaseModel):
    currency: str
    list_price: float
    sale_price: float
    discount_reason: str | None = None


class OfferFeatureFlagsResponse(BaseModel):
    delivery_mode: str
    teacher_included: bool
    homework_review_included: bool


class BundleComponentResponse(BaseModel):
    offer_id: str
    quantity: int
    position: int


class InternalOfferSnapshotResponse(BaseModel):
    offer_id: str
    course_id: str
    offer_code: str
    title: str
    is_active: bool
    price: MoneyResponse
    feature_flags: OfferFeatureFlagsResponse


class InternalBundleSnapshotResponse(BaseModel):
    bundle_offer_id: str
    title: str
    is_active: bool
    price: MoneyResponse
    components: list[BundleComponentResponse]